+++ kitchen-sink/app.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import CSSTransitionGroup from 'react-addons-css-transition-group';
 import {
   Router,
   Route,
   Link,
-  History,
+  hashHistory,
   IndexRoute,
 } from 'react-router';
 
@@ -79,10 +78,6 @@ const NotFound = React.createClass({
 });
 
 const Detail = React.createClass({
-  // DO NOT NEED HERE.
-  // https://github.com/rackt/react-router/blob/master/docs/API.md#history-mixin
-  // mixins: [History],
-
   render() {
     let component = this.props.params.component;
 
@@ -95,10 +90,7 @@ const Detail = React.createClass({
       component: Link,
       icon: 'left-nav',
       title: '返回',
-      props: {
-        to: '/',
-        // onClick: () => this.props.history.goBack(),
-      }
+      to: '/',
     };
 
     return (
@@ -117,7 +109,7 @@ const Detail = React.createClass({
 });
 
 const routes = (
-  <Router>
+  <Router history={hashHistory}>
     <Route path="/" component={App}>
       <Route path=":component" component={Detail} />
       <IndexRoute component={Default} />