+++ src/js/OffCanvasTrigger.js
@@ -1,6 +1,10 @@
-import React, {cloneElement} from 'react';
-import ReactDOM from 'react-dom';
-import classNames from 'classnames';
+import React, {
+  cloneElement,
+  PropTypes,
+} from 'react';
+import {
+  findDOMNode,
+} from 'react-dom';
 import ClassNameMixin from './mixins/ClassNameMixin';
 import OverlayMixin from './mixins/OverlayMixin';
 import CSSCore from './utils/CSSCore';
@@ -11,13 +15,13 @@ const OffCanvasTrigger = React.createClass({
   mixins: [OverlayMixin, ClassNameMixin],
 
   propTypes: {
-    defaultOffCanvasActive: React.PropTypes.bool,
-    placement: React.PropTypes.oneOf(['left', 'right']),
-    animation: React.PropTypes.oneOf(['slide', 'push']),
-    offCanvas: React.PropTypes.node.isRequired,
-    pageContainer: React.PropTypes.node,
-    onOpen: React.PropTypes.func,
-    onClosed: React.PropTypes.func,
+    defaultOffCanvasActive: PropTypes.bool,
+    placement: PropTypes.oneOf(['left', 'right']),
+    animation: PropTypes.oneOf(['slide', 'push']),
+    offCanvas: PropTypes.node.isRequired,
+    pageContainer: PropTypes.node,
+    onOpen: PropTypes.func,
+    onClosed: PropTypes.func,
   },
 
   getDefaultProps() {
@@ -40,10 +44,6 @@ const OffCanvasTrigger = React.createClass({
   },
 
   componentDidMount() {
-    if (this.props.defaultOffCanvasActive) {
-      this.updateOffCanvasPosition();
-    }
-
     this.setPageContainer();
   },
 
@@ -110,11 +110,10 @@ const OffCanvasTrigger = React.createClass({
   getPageContainer() {
     let {
       pageContainer
-      } = this.props;
+    } = this.props;
 
     return typeof pageContainer === 'string' ?
-      document.querySelector(pageContainer) :
-      ReactDOM.findDOMNode(pageContainer);
+      document.querySelector(pageContainer) : findDOMNode(pageContainer);
   },
 
   setPageContainer() {
@@ -142,7 +141,7 @@ const OffCanvasTrigger = React.createClass({
     let offCanvas = this.props.offCanvas;
     let {
       isClosing,
-      } = this.state;
+    } = this.state;
 
     if (isClosing) {
       let node = this.getOverlayDOMNode();